pi/scripts/blink_test_requirements.txt
# For Pi Zero 2 blink test: pip install -r scripts/blink_test_requirements.txt
# RPi.GPIO comes with Raspberry Pi OS
pyserial>=3.5
